refactor the refactor
